+++ custom-llvm/prepare.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#===-- build.bash -------*- bash script -*-===//
+#===-- build.bash ======-*- bash script -*-===//
 #
 # This file is distributed under the GPL v3 license (https://www.gnu.org/licenses/gpl-3.0.en.html).
 #
@@ -20,31 +20,40 @@ echo "************************************************"
 echo "* This script helps to download source code of *"
 echo "*            clang, llvm, compiler-rt          *"
 echo "************************************************"
-
-set -x 
+echo 
+echo "Futag will collect information for preparing build system."
+echo "========================================================="
+echo "-- [Futag] Select version of llvm for building:"
+echo "-- 1. LLVM 14.0.6"
+echo "-- 2. LLVM 13.0.1"
+echo "-- 3. LLVM 12.0.1"
+llvmVersion=(1 2 3)
+read -p "-- Your choice (1/2/3 - default to 1): " selectedVersion 
+if [[ ! " ${llvmVersion[*]} " =~ " ${selectedVersion} " ]]; then
+echo "-- [Futag] Wrong input! Please enter 1, 2 or 3! Exit..."
+exit
+fi
+echo
+# https://github.com/AFLplusplus/AFLplusplus/archive/refs/tags/4.02c.tar.gz
+echo "========================================================="
+read -p "-- [Futag] Build with AFLplusplus-4.02c? (y/n): " wAFLplusplus
+if [[ ! $wAFLplusplus == [yYnN] ]]; then
+    echo "-- [Futag] Wrong input! Please enter y or n! Exit..."
+    exit
+fi
+# https://github.com/ossf/fuzz-introspector/archive/refs/tags/v1.0.0.tar.gz
+echo "========================================================="
+read -p "-- [Futag] Build with fuzz-introspector? (y/n): " fuzzintro
+if [[ ! $fuzzintro == [yYnN] ]]; then
+    echo "-- [Futag] Wrong input! Please enter y or n! Exit..."
+    exit
+fi
+echo "========================================================="
+echo
+echo "-- [Futag] Preparing .. "
 
 futag_install_folder="$(pwd)/../futag-llvm"
 build_folder="$(pwd)/../build"
-
-if [ -d "AFLplusplus-4.02c" ]
-then
-    rm -rf AFLplusplus-4.02c
-fi
-
-if [ -d "fuzz-introspector" ]
-then
-    rm -rf fuzz-introspector
-fi
-
-if [ -f llvm-project-14.0.6.src.tar.xz ]
-then
-    rm llvm-project-14.0.6.src.tar.xz
-fi
-
-if [ -d "llvm-project-14.0.6.src" ]
-then
-    rm -rf llvm-project-14.0.6.src
-fi
 
 #create build folder and copy script
 if [ -d $build_folder ]
@@ -53,11 +62,6 @@ then
 fi
 mkdir $build_folder
 
-#getting llvm-project 
-wget https://github.com/llvm/llvm-project/releases/download/llvmorg-14.0.6/llvm-project-14.0.6.src.tar.xz
-
-for f in *.tar.*; do tar xf "$f"; done
-
 # create futag installation folder
 if [ -d $futag_install_folder ]
 then
@@ -65,13 +69,72 @@ then
 fi
 mkdir $futag_install_folder
 
-# begin integrate with fuzz-introspector
-# extract and build binutils
-binutils_build="binutils-build"
+#write infomation of build to file INFO
+file_info="INFO"
+set -x
+if [ -d "llvm-project" ]; then
+    rm -rf llvm-project
+fi
+if [ "$selectedVersion" == "1" ]; then
+    echo "LLVM=14.0.6" > $file_info
+    if [ ! -f llvm-project-14.0.6.src.tar.xz ]; then
+        wget https://github.com/llvm/llvm-project/releases/download/llvmorg-14.0.6/llvm-project-14.0.6.src.tar.xz
+    fi
+    tar xf llvm-project-14.0.6.src.tar.xz
+    mv llvm-project-14.0.6.src llvm-project
+fi
 
-mv binutils $build_folder/
-# git clone --depth 1 git://sourceware.org/git/binutils-gdb.git binutils
+if [ "$selectedVersion" == "2" ]; then
+    echo "LLVM=13.0.1" > $file_info
+    if [ ! -f llvm-project-13.0.1.src.tar.xz ]; then
+        wget https://github.com/llvm/llvm-project/releases/download/llvmorg-13.0.1/llvm-project-13.0.1.src.tar.xz
+    fi
+    tar xf llvm-project-13.0.1.src.tar.xz
+    mv llvm-project-13.0.1.src llvm-project
+fi
 
-set +x 
-mv fuzz-introspector $futag_install_folder/
-cp build*.sh  $build_folder
+if [ "$selectedVersion" == "3" ]; then
+    echo "LLVM=12.0.1" > $file_info
+    if [ ! -f llvm-project-12.0.1.src.tar.xz ]; then
+        wget https://github.com/llvm/llvm-project/releases/download/llvmorg-12.0.1/llvm-project-12.0.1.src.tar.xz
+    fi
+    tar xf llvm-project-12.0.1.src.tar.xz
+    mv llvm-project-12.0.1.src llvm-project
+fi
+build_script="build.sh"
+
+if [ $wAFLplusplus == "Y" ] || [ $wAFLplusplus == "y" ]; then
+    echo "AFLplusplus=yes" >> $file_info
+    if [ -d AFLplusplus-4.02c ]; then
+        rm -rf AFLplusplus-4.02c
+    fi
+    if [ ! -f 4.02c.tar.gz ]; then
+        wget https://github.com/AFLplusplus/AFLplusplus/archive/refs/tags/4.02c.tar.gz
+    fi
+    tar xf 4.02c.tar.gz
+    mv AFLplusplus-4.02c $build_folder/
+    build_script="buildwAFLplusplus.sh"
+else
+    echo "AFLplusplus=no" >> $file_info
+fi
+
+if [ $fuzzintro == "Y" ] || [ $fuzzintro == "y" ]; then
+    echo "FuzzIntrospector=yes" >> $file_info
+    if [ -d fuzz-introspector-1.0.0 ]; then
+        rm -rf fuzz-introspector-1.0.0
+    fi
+    if [ ! -f v1.0.0.tar.gz ]; then
+        wget https://github.com/ossf/fuzz-introspector/archive/refs/tags/v1.0.0.tar.gz
+    fi
+    tar xf v1.0.0.tar.gz
+    cp -r fuzz-introspector-1.0.0 $futag_install_folder/
+    if [ $wAFLplusplus == "Y" ] || [ $wAFLplusplus == "y" ]; then
+        build_script="buildwAFLplusplusFuzzIntro.sh"
+    else
+        build_script="buildwFuzzIntro.sh"
+    fi
+else
+    echo "FuzzIntrospector=no" >> $file_info
+fi
+git rev-parse HEAD >> $file_info
+cp $build_script  $build_folder/build.sh