+++ src/python/futag-package/src/futag/preprocessor.py
@@ -19,7 +19,7 @@ import os
 import re
 
 from futag.sysmsg import *
-from subprocess import Popen, PIPE
+from subprocess import Popen, PIPE, run
 
 def delete_folder(pth):
     """
@@ -191,9 +191,9 @@ class Builder:
             "-disable-checker",
             "cplusplus",
             "-enable-checker",
-            "futag.FutagFunctionAnalyzer",
+            "futag.FutagAnalyzer",
             "-analyzer-config",
-            "futag.FutagFunctionAnalyzer:report_dir=" + self.analysis_path.as_posix(),
+            "futag.FutagAnalyzer:report_dir=" + self.analysis_path.as_posix(),
             "make",
             "-j" + str(self.processes)
         ], stdout=PIPE, stderr=PIPE, universal_newlines=True, env=my_env)
@@ -243,6 +243,7 @@ class Builder:
         # Doing make for building
 
         p = Popen([
+            (self.futag_llvm_package / "bin/intercept-build").as_posix(),
             "make",
             "-j" + str(self.processes)
         ], stdout=PIPE, stderr=PIPE, universal_newlines=True, env=my_env)
@@ -324,9 +325,9 @@ class Builder:
             "-disable-checker",
             "cplusplus",
             "-enable-checker",
-            "futag.FutagFunctionAnalyzer",
+            "futag.FutagAnalyzer",
             "-analyzer-config",
-            "futag.FutagFunctionAnalyzer:report_dir=" + self.analysis_path.as_posix(),
+            "futag.FutagAnalyzer:report_dir=" + self.analysis_path.as_posix(),
             "make",
             "-j" + str(self.processes)
         ], stdout=PIPE, stderr=PIPE, universal_newlines=True)
@@ -373,6 +374,7 @@ class Builder:
             print(LIB_CONFIGURE_SUCCEEDED)
 
         p = Popen([
+            (self.futag_llvm_package / "bin/intercept-build").as_posix(),
             "make",
             "-j" + str(self.processes)
         ], stdout=PIPE, stderr=PIPE, universal_newlines=True, env=my_env)
@@ -428,9 +430,9 @@ class Builder:
             "-disable-checker",
             "cplusplus",
             "-enable-checker",
-            "futag.FutagFunctionAnalyzer",
+            "futag.FutagAnalyzer",
             "-analyzer-config",
-            "futag.FutagFunctionAnalyzer:report_dir=" + self.analysis_path.as_posix(),
+            "futag.FutagAnalyzer:report_dir=" + self.analysis_path.as_posix(),
             "make",
             "-j" + str(self.processes)
         ], stdout=PIPE, stderr=PIPE, universal_newlines=True)
@@ -459,6 +461,7 @@ class Builder:
         my_env["CXX"] = (self.futag_llvm_package / 'bin/clang++').as_posix()
 
         p = Popen([
+            (self.futag_llvm_package / "bin/intercept-build").as_posix(),
             "make",
             "-j" + str(self.processes)
         # ], stdout=PIPE, stderr=PIPE, universal_newlines=True, env=my_env)
@@ -497,26 +500,26 @@ class Builder:
         """
         decl_files = [
             x
-            for x in self.analysis_path.glob("**/declaration-*.futag-function-analyzer")
+            for x in self.analysis_path.glob("**/declaration-*.futag-analyzer.json")
             if x.is_file()
         ]
 
-        # Find all context files in given location
-        context_files = [
-            x for x in self.analysis_path.glob("**/context-*.futag-function-analyzer") if x.is_file()
-        ]
+        # # Find all context files in given location
+        # context_files = [
+        #     x for x in self.analysis_path.glob("**/context-*.futag-function-analyzer.json") if x.is_file()
+        # ]
 
         # Find all type_info files in given location
         typeinfo_files = [
             x
-            for x in self.analysis_path.glob("**/types-info-*.futag-function-analyzer")
+            for x in self.analysis_path.glob("**/types-info-*.futag-analyzer.json")
             if x.is_file()
         ]
 
         # Find all includes info files in given location
         info_files = [
             x
-            for x in self.analysis_path.glob("**/file-info-*.futag-function-analyzer")
+            for x in self.analysis_path.glob("**/file-info-*.futag-analyzer.json")
             if x.is_file()
         ]
 
@@ -525,49 +528,57 @@ class Builder:
         enum_list = []
         typedef_list = []
         struct_list = []
+        class_list = []
+        union_list = []
+        unknown_record_list = []
         compiled_files = []
 
         print("")
         print(" -- [Futag]: Analysing fuctions for generating fuzz-drivers..." )
+        print (decl_files)
         for jf in decl_files:
             functions = json.load(open(jf.as_posix()))
             if functions is None:
                 print(" -- [Futag]: Warning: loading json from file %s failed!" %
                       (jf.as_posix()))
                 continue
+            else:
+                print(" -- [Futag]: Analyzing file %s ..." %
+                      (jf.as_posix()))
             # get global hash of all functions
             global_hash = [x for x in function_list]
             # iterate function hash for adding to global hash list
             for hash in functions:
+                curr_func = functions[hash]
                 if not hash in global_hash:
                     function_list[hash] = functions[hash]
 
-        for jf in context_files:
-            contexts = json.load(open(jf.as_posix()))
-            if contexts is None:
-                print(" -- [Futag]: Warning: loading json from file %s failed!" %
-                      (jf.as_posix()))
-                continue
-            # get global hash of all functions
-            global_hash = [x for x in function_list]
+        # for jf in context_files:
+        #     contexts = json.load(open(jf.as_posix()))
+        #     if contexts is None:
+        #         print(" -- [Futag]: Warning: loading json from file %s failed!" %
+        #               (jf.as_posix()))
+        #         continue
+        #     # get global hash of all functions
+        #     global_hash = [x for x in function_list]
 
-            # iterate function hash for adding to global hash list
-            for hash in contexts:
-                if hash in global_hash:
-                    called_from_list = [
-                        x["called_from"] + x["called_from_func_name"]
-                        for x in function_list[hash]["call_contexts"]
-                    ]
-                    for call_xref in contexts[hash]["call_contexts"]:
-                        if (
-                            not call_xref["called_from"] +
-                                call_xref["called_from_func_name"]
-                            in called_from_list
-                        ):
-                            function_list[hash]["call_contexts"].append(
-                                call_xref)
-                else:
-                    print(" -- %s not found in global hash list!" % (hash))
+        #     # iterate function hash for adding to global hash list
+        #     for hash in contexts:
+        #         if hash in global_hash:
+        #             called_from_list = [
+        #                 x["called_from"] + x["called_from_func_name"]
+        #                 for x in function_list[hash]["call_contexts"]
+        #             ]
+        #             for call_xref in contexts[hash]["call_contexts"]:
+        #                 if (
+        #                     not call_xref["called_from"] +
+        #                         call_xref["called_from_func_name"]
+        #                     in called_from_list
+        #                 ):
+        #                     function_list[hash]["call_contexts"].append(
+        #                         call_xref)
+        #         else:
+        #             print(" -- %s not found in global hash list!" % (hash))
 
         print("")
         print(" -- [Futag]: Analysing data types ..." )
@@ -578,6 +589,8 @@ class Builder:
                 print(" -- [Futag]: Warning: loading json from file %s failed!" %
                       (jf.as_posix()))
                 continue
+            else:
+                print(" -- [Futag]: Analyzing file %s ..." % (jf.as_posix()))
             # get global hash of all functions
             for enum_it in types["enums"]:
                 exist = False
@@ -588,16 +601,48 @@ class Builder:
                 if not exist:
                     enum_list.append(enum_it)
 
-            for struct_it in types["structs"]:
-                exist = False
-                for struct_exist_it in struct_list:
-                    if struct_it["struct_name"] == struct_exist_it["struct_name"]:
-                        if len(struct_it["struct_fields"]) > len(struct_exist_it["struct_fields"]):
-                            struct_exist_it["struct_fields"] = struct_it["struct_fields"]
-                        exist = True
-                        break
-                if not exist:
-                    struct_list.append(struct_it)
+            for record in types["records"]:
+                if record["type"] == STRUCT_RECORD:
+                    exist = False
+                    for struct_iter in struct_list:
+                        if record["name"] == struct_iter["name"]:
+                            if len(record["fields"]) > len(struct_iter["fields"]):
+                                struct_iter["fields"] = record["fields"]
+                            exist = True
+                            break
+                    if not exist:
+                        struct_list.append(record)
+                if record["type"] == UNION_RECORD :
+                    exist = False
+                    for union_iter in union_list:
+                        if record["name"] == union_iter["name"]:
+                            if len(record["fields"]) > len(union_iter["fields"]):
+                                union_iter["fields"] = record["fields"]
+                            exist = True
+                            break
+                    if not exist:
+                        union_list.append(record)
+                if record["type"] == CLASS_RECORD :
+                    exist = False
+                    for class_iter in class_list:
+                        if record["name"] == class_iter["name"]:
+                            if len(record["fields"]) > len(class_iter["fields"]):
+                                class_iter["fields"] = record["fields"]
+                            exist = True
+                            break
+                    if not exist:
+                        class_list.append(record)
+                if record["type"] == UNKNOW_RECORD :
+                    exist = False
+                    for record_iter in unknown_record_list:
+                        if record["name"] == record_iter["name"]:
+                            if len(record["fields"]) > len(record_iter["fields"]):
+                                record_iter["fields"] = record["fields"]
+                            exist = True
+                            break
+                    if not exist:
+                        unknown_record_list.append(record)
+                
 
             for typedef_it in types["typedefs"]:
                 exist = False
@@ -646,6 +691,9 @@ class Builder:
                         contexts.append(call_func)
             fs = {
                 "func_name": function_list[func]["func_name"],
+                "func_qname": function_list[func]["func_qname"],
+                "vision": function_list[func]["vision"],
+                "parent_name": function_list[func]["parent_name"],
                 "return_type": function_list[func]["return_type"],
                 "return_type_pointer": function_list[func]["return_type_pointer"],
                 "params": function_list[func]["params"],
@@ -660,6 +708,9 @@ class Builder:
             "functions": functions_w_contexts,
             "enums": enum_list,
             "structs": struct_list,
+            "unions": union_list,
+            "classes": class_list,
+            "unknown_records": unknown_record_list,
             "typedefs": typedef_list,
             "compiled_files": compiled_files,
         }
@@ -669,6 +720,9 @@ class Builder:
         print("Total functions: ", str(len(result["functions"])))
         print("Total enums: ", str(len(result["enums"])))
         print("Total structs: ", str(len(result["structs"])))
+        print("Total unions: ", str(len(result["unions"])))
+        print("Total classes: ", str(len(result["classes"])))
+        print("Total unknown records: ", str(len(result["unknown_records"])))
         print("Total typedefs: ", str(len(result["typedefs"])))
         print("Analysis result: ", (self.analysis_path /
               "futag-analysis-result.json").as_posix())