+++ src/python/template-script.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python3
 from futag.preprocessor import *
+from futag.generator import * 
 
 lib_test = Builder(
 "../futag-llvm",        #Путь к рабочей директории futag
 "../json-c",                    #Путь к директории исходных текстов исследуемого приложения
+"-g -O0",                 #Флаги при сборке
 True,                           #Очистить каталоги futag-build, futag-install, futag-analysis перед запуском, допустимые значение: (True/False)(Необязательный параметр, по-умолчанию False)
 "../json-c/futag-build",        #Путь к директории futag-build (Необязательный параметр)
 "../json-c/futag-install",      #Путь к директории futag-install (Необязательный параметр)
@@ -18,8 +20,8 @@ lib_test.analyze()
 lib_test = Generator(
     "../futag-llvm/",
     "json-c",
-    )
+)
 lib_test.gen_targets()
-lib_test.compile_targets()
+lib_test.compile_targets(True, 4)
 
 print("-- [Futag]: fuzz-drivers are saved in json-c/futag-fuzz-targets!")